backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
gigachat